% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_data.R
\name{get_example_data}
\alias{get_example_data}
\title{Get example data}
\usage{
get_example_data(dataset = c("MIsim", "relhaz", "frailty"),
  where = ".")
}
\arguments{
\item{dataset}{The dataset to copy. Possible choices are: \code{MIsim}, \code{relhaz}, \code{frailty}. More information on each dataset can be obtained by typing \code{?interest::MIsim}, \code{?interest::relhaz}, or \code{?interest::frailty} in the R console.}

\item{where}{Where to copy the dataset. Defaults to the current directory.}
}
\value{
The required dataset, invisibly.
}
\description{
Copy an example dataset to use with INTEREST to a given directory.
}
\examples{
\dontrun{
get_example_data("MIsim")
}
}
