\name{interest-package}
\alias{interest-package}
\alias{interest}
\docType{package}
\title{
\packageTitle{interest}
}
\description{
\packageDescription{interest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{interest}
\packageIndices{interest}
}
\author{
\packageAuthor{interest}
Maintainer: Peter Gruber
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
