\name{mdsChr22Obj}
\alias{mdsChr22Obj}
\docType{data}
\title{
	Object of InterestResults type for MDS data
}
\description{
	The Results of IntEREst analysis (with Intron retention levels) for the genes that feature U12-type introns and are located on Chr22 in MDS data.
}
\usage{data(mdsChr22Obj)}
\format{
  An Object of class \code{InterestResults} that contains intron retention results generated by \code{interest()} function on MDS data consisting of bone-marrows samples of 8 MDS patients with ZRSR2 mutations, 4 patients without the mutation and 4 healthy individuals.
  \describe{
    \item{\code{@resultFiles}}{Vector containing paths to the result files.}
    \item{\code{@readFreq}}{Numeric matrix that includes the raw read counts.}
    \item{\code{@scaledRetention}}{Numeric matrix that includes the scaled retention values.}
    \item{\code{@sampleNames}}{Name of the samples (Vector of characters).}
	\item{\code{@scaleLength}}{
	Logical value, indicating whether the intron/exon retention levels are scaled to the length of the introns/exons. 
	}
	\item{\code{@scaleFragment}}{
	Logical value, indicating whether the intron/exon retention levels are scaled to the fragments mapped to the genes. 
	}
    \item{\code{@sampleAnnotation}}{A data frame that its first column is the sample names and the other columns are annotations of the samples. The column names are characters that describe the annotations.}
	\item{\code{@interestDf}}{A data.frame that features intron/exon annotations and their results of IntEREst analysis, e.g. row and normalized retention levels.}
  }
}
\value{
Object of class \code{InterestResult}. 
}
\source{
Madan, V., et.al.,  Aberrant splicing of U12-type introns is the hallmark of ZRSR2 mutant myelodysplastic syndrome. Nat Communication 2015 Jan 14;6:6042. doi: 10.1038/ncomms7042.
}
\keyword{datasets}
