% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plots a time series}
\usage{
plot_time_series(data)
}
\arguments{
\item{data}{A data frame with \code{time}, \code{value} and \code{id} columns.}
}
\description{
Creates a plot for a ts-boxable dataset.
}
