% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_theme.R
\name{apply_theme}
\alias{apply_theme}
\title{Apply theme to a ggplot object}
\usage{
apply_theme(ggobject, theme)
}
\arguments{
\item{ggobject}{A \code{ggplot2} object.}

\item{theme}{A string specifying the theme to be added to the ggplot2 plot. Possible choices are: \code{theme_bw}, \code{theme_dark}, \code{theme_few}, \code{theme_grey}, \code{theme_light}, \code{theme_linedraw}, \code{theme_minimal}, \code{theme_tufte}, \code{theme_viridis}. Themes, color scales, and fill scales are imported from the \code{ggplot2}, \code{ggthemes}, \code{viridis}}
}
\value{
A \code{ggplot2} object.
}
\description{
Apply theme to a ggplot object
}
