\name{IntEREst-package}
\alias{IntEREst-package}
\alias{IntEREst}
\docType{package}
\title{
IntEREst
}
\description{
Intron/Exon retention estimator quantifies and normalizes Intron retention and 
Exon junction read levels by analyzing mapped reads (.bam) files.
}
\details{
\tabular{ll}{
Package: \tab IntEREst\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-11-18\cr
License: GPL-2
}
To run the pipeline use functions \code{interest()} or 
\code{interest.sequential()}, i.e. wrapper functions that run all the necessary
functions.
}
\author{
Ali Oghabian <Ali.Oghabian@Helsinki.Fi>, 
Dario Greco <dario.greco@helsinki.fi>, 
Mikko Frilander <Mikko.Frilander@helsinki.fi>

Maintainer: Ali Oghabian <Ali.Oghabian@Helsinki.Fi>,
Mikko Frilander <Mikko.Frilander@helsinki.fi>
}

\keyword{package} 
\keyword{expression} 
\keyword{sequencing}
\keyword{splicing}
\keyword{rna-seq} 
\keyword{intron} 
\keyword{retention}
