% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest.R
\name{interest}
\alias{interest}
\title{INTEREST}
\usage{
interest(...)
}
\arguments{
\item{...}{Passed onto \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\description{
Run the INTEREST interactive shiny app.
}
\details{
interest
}
\examples{
interest()
}
